Refactor eventing api group lifecycle hooks
Fixes https://github.com/knative/eventing/issues/980

Refactor trigger, subscription, broker and channel API groups to have
lifecycle hooks

- no feature or new code introduced in this PR
