+++ pkg/apis/eventing/v1alpha1/channel_types.go
@@ -77,8 +77,6 @@ type ChannelSpec struct {
 	Subscribable *eventingduck.Subscribable `json:"subscribable,omitempty"`
 }
 
-var chanCondSet = duckv1alpha1.NewLivingConditionSet(ChannelConditionProvisioned, ChannelConditionAddressable, ChannelConditionProvisionerInstalled)
-
 // ChannelStatus represents the current state of a Channel.
 type ChannelStatus struct {
 	// inherits duck/v1alpha1 Status, which currently provides:
@@ -96,75 +94,6 @@ type ChannelStatus struct {
 	// Internal is status unique to each ClusterChannelProvisioner.
 	// +optional
 	Internal *runtime.RawExtension `json:"internal,omitempty"`
-}
-
-const (
-	// ChannelConditionReady has status True when the Channel is ready to
-	// accept traffic.
-	ChannelConditionReady = duckv1alpha1.ConditionReady
-
-	// ChannelConditionProvisioned has status True when the Channel's
-	// backing resources have been provisioned.
-	ChannelConditionProvisioned duckv1alpha1.ConditionType = "Provisioned"
-
-	// ChannelConditionAddressable has status true when this Channel meets
-	// the Addressable contract and has a non-empty hostname.
-	ChannelConditionAddressable duckv1alpha1.ConditionType = "Addressable"
-
-	// ChannelConditionProvisionerFound has status true when the channel is being watched
-	// by the provisioner's channel controller (in other words, the provisioner is installed)
-	ChannelConditionProvisionerInstalled duckv1alpha1.ConditionType = "ProvisionerInstalled"
-)
-
-// GetCondition returns the condition currently associated with the given type, or nil.
-func (cs *ChannelStatus) GetCondition(t duckv1alpha1.ConditionType) *duckv1alpha1.Condition {
-	return chanCondSet.Manage(cs).GetCondition(t)
-}
-
-// IsReady returns true if the resource is ready overall.
-func (cs *ChannelStatus) IsReady() bool {
-	return chanCondSet.Manage(cs).IsHappy()
-}
-
-// InitializeConditions sets relevant unset conditions to Unknown state.
-func (cs *ChannelStatus) InitializeConditions() {
-	chanCondSet.Manage(cs).InitializeConditions()
-	// Channel-default-controller sets ChannelConditionProvisionerInstalled=False, and it needs to be set to True by individual controllers
-	// This is done so that each individual channel controller gets it for free.
-	// It is also implied here that the channel-default-controller never calls InitializeConditions(), while individual channel controllers
-	// call InitializeConditions() as one of the first things in its reconcile loop.
-	cs.MarkProvisionerInstalled()
-}
-
-// MarkProvisioned sets ChannelConditionProvisioned condition to True state.
-func (cs *ChannelStatus) MarkProvisioned() {
-	chanCondSet.Manage(cs).MarkTrue(ChannelConditionProvisioned)
-}
-
-// MarkNotProvisioned sets ChannelConditionProvisioned condition to False state.
-func (cs *ChannelStatus) MarkNotProvisioned(reason, messageFormat string, messageA ...interface{}) {
-	chanCondSet.Manage(cs).MarkFalse(ChannelConditionProvisioned, reason, messageFormat, messageA...)
-}
-
-// MarkProvisionerInstalled sets ChannelConditionProvisionerInstalled condition to True state.
-func (cs *ChannelStatus) MarkProvisionerInstalled() {
-	chanCondSet.Manage(cs).MarkTrue(ChannelConditionProvisionerInstalled)
-}
-
-// MarkProvisionerNotInstalled sets ChannelConditionProvisionerInstalled condition to False state.
-func (cs *ChannelStatus) MarkProvisionerNotInstalled(reason, messageFormat string, messageA ...interface{}) {
-	chanCondSet.Manage(cs).MarkFalse(ChannelConditionProvisionerInstalled, reason, messageFormat, messageA...)
-}
-
-// SetAddress makes this Channel addressable by setting the hostname. It also
-// sets the ChannelConditionAddressable to true.
-func (cs *ChannelStatus) SetAddress(hostname string) {
-	cs.Address.Hostname = hostname
-	if hostname != "" {
-		chanCondSet.Manage(cs).MarkTrue(ChannelConditionAddressable)
-	} else {
-		chanCondSet.Manage(cs).MarkFalse(ChannelConditionAddressable, "emptyHostname", "hostname is the empty string")
-	}
 }
 
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
